decouple mmr-primitives
