+++ modules/ismp/pallets/pallet/Cargo.toml
@@ -25,7 +25,6 @@ ismp = { workspace = true }
 codec = { workspace = true }
 scale-info = { workspace = true }
 sp-mmr-primitives = { workspace = true }
-mmr-primitives = { workspace = true }
 anyhow = { workspace = true, default-features = false }
 
 # crates.io
@@ -55,7 +54,6 @@ std = [
     "sp-api/std",
     "serde/default",
     "sp-mmr-primitives/std",
-    "mmr-primitives/std",
     "anyhow/std"
 ]
 disable-panic-handler = ["sp-io/disable_panic_handler", "sp-io/disable_oom", "sp-io/disable_allocator"]
