+++ modules/trees/mmr/primitives/Cargo.toml
@@ -10,7 +10,7 @@ documentation = "https://docs.rs/ismp-parachain"
 description = "Merkle mountain range primitives for pallet-ismp"
 keywords = ["substrate", "polkadot-sdk", "ISMP", "interoperability"]
 readme = "./README.md"
-
+publish = false
 
 [dependencies]
 merkle-mountain-range = { workspace = true }
@@ -19,10 +19,11 @@ sp-core = { workspace = true }
 sp-runtime = { workspace = true }
 sp-std = { workspace = true }
 sp-mmr-primitives = { workspace = true }
-serde = { workspace = true, features = ["derive"], default-features = false }
+serde = { workspace = true, features = ["derive"] }
 scale-info = { workspace = true }
 frame-system = { workspace = true }
 ismp = { workspace = true }
+pallet-ismp = { workspace = true }
 sp-io = { workspace = true }
 
 [features]
@@ -38,5 +39,6 @@ std = [
     "merkle-mountain-range/std",
     "frame-system/std",
     "ismp/std",
+    "pallet-ismp/std",
     "sp-io/std"
 ]
